@@ -23,7 +23,6 @@ buildscript {
 subprojects {
     apply plugin: 'java'
     apply plugin: 'jacoco'
-    apply plugin: 'org.openjfx.javafxplugin'
 
     apply plugin: 'idea'
     apply plugin: 'com.github.ben-manes.versions'
@@ -31,19 +30,35 @@ subprojects {
     group = 'se.llbit'
     version = rootProject.version
 
-    sourceCompatibility = '1.8'
-    targetCompatibility = '1.8'
-
-    compileJava {
-        options.encoding = 'UTF-8'
-        options.debug = true
-        options.deprecation = true
+    java {
+        toolchain {
+            // default java version for the project, some subprojects are pinned to java 8 for backwards compatibility (see their configs).
+            languageVersion = JavaLanguageVersion.of(17)
+        }
     }
+    compileTestJava {
+        // compile tests using more recent features
+        javaCompiler = javaToolchains.compilerFor {
+            languageVersion = JavaLanguageVersion.of(21)
+        }
+    }
+    test {
+        // run tests using more recent features
+        javaLauncher = javaToolchains.launcherFor {
+            languageVersion = JavaLanguageVersion.of(21)
+        }
+        useJUnitPlatform()
 
-    javafx {
-        version = '11.0.2'
-        configuration = 'implementation'
-        modules = ['javafx.base', 'javafx.controls', 'javafx.fxml']
+        // Always run tests, even when nothing changed.
+        dependsOn 'cleanTest'
+
+        finalizedBy jacocoTestReport // report is always generated after tests run (it's very cheap)
+    }
+    jacoco {
+        toolVersion = "0.8.11"
+    }
+    jacocoTestReport {
+        dependsOn test // tests are required to run before generating the report
     }
 
     javadoc {
@@ -63,6 +78,12 @@ subprojects {
         nbtlib 'se.llbit:jo-nbt:1.3.0'
         cplib 'se.llbit:luxcp:1.0.1'
         toolpanelib 'se.llbit:toolpane:0.1'
+
+        testImplementation("org.assertj:assertj-core:3.25.1")
+
+        testImplementation(platform('org.junit:junit-bom:5.10.2'))
+        testImplementation('org.junit.jupiter:junit-jupiter')
+        testRuntimeOnly('org.junit.platform:junit-platform-launcher')
     }
 
     idea {
@@ -236,6 +257,7 @@ task docs(type: Javadoc) {
 }
 
 task clean {
+    group = "Build"
     doLast {
         delete "./build"
     }
