@@ -17,7 +17,7 @@
  */
 package se.llbit.chunky.renderer.scene;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class SceneTest {
   /**
