@@ -16,9 +16,9 @@
  */
 package se.llbit.chunky.resources;
 
-import static org.junit.Assert.assertEquals;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class AnimatedTextureTest {
 
