@@ -18,13 +18,14 @@ package se.llbit.fxutil;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class GroupedChangeListenerTest {
-  @Test public void testRecursive() {
+  @Test
+  public void testRecursive() {
     BooleanProperty p1 = new SimpleBooleanProperty(false);
     BooleanProperty p2 = new SimpleBooleanProperty(false);
     BooleanProperty p3 = new SimpleBooleanProperty(false);
