@@ -16,9 +16,10 @@
  */
 package se.llbit.math;
 
-import static org.junit.Assert.assertTrue;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * TODO add more tests
@@ -33,10 +34,10 @@ public class QuickMathTest {
 	 */
 	@Test
 	public void testMin_1() {
-		assertTrue(11.0 == QuickMath.min(132.0, 11.0));
-		assertTrue(11.0 == QuickMath.min(11.0, 132.0));
-		assertTrue(-132.0 == QuickMath.min(-132.0, -11.0));
-		assertTrue(-132.0 == QuickMath.min(-11.0, -132.0));
+    assertEquals(11.0, QuickMath.min(132.0, 11.0));
+    assertEquals(11.0, QuickMath.min(11.0, 132.0));
+    assertEquals(-132.0, QuickMath.min(-132.0, -11.0));
+    assertEquals(-132.0, QuickMath.min(-11.0, -132.0));
 	}
 
 	/**
@@ -44,7 +45,7 @@ public class QuickMathTest {
 	 */
 	@Test
 	public void testMin_2() {
-		assertTrue(11.0 == QuickMath.min(Double.NaN, 11.0));
+    assertEquals(11.0, QuickMath.min(Double.NaN, 11.0));
 		assertTrue(Double.isNaN(QuickMath.min(11.0, Double.NaN)));
 	}
 
@@ -53,10 +54,10 @@ public class QuickMathTest {
 	 */
 	@Test
 	public void testMax_1() {
-		assertTrue(132.0 == QuickMath.max(132.0, 11.0));
-		assertTrue(132.0 == QuickMath.max(11.0, 132.0));
-		assertTrue(-11.0 == QuickMath.max(-132.0, -11.0));
-		assertTrue(-11.0 == QuickMath.max(-11.0, -132.0));
+    assertEquals(132.0, QuickMath.max(132.0, 11.0));
+    assertEquals(132.0, QuickMath.max(11.0, 132.0));
+    assertEquals(-11.0, QuickMath.max(-132.0, -11.0));
+    assertEquals(-11.0, QuickMath.max(-11.0, -132.0));
 	}
 
 	/**
@@ -64,7 +65,7 @@ public class QuickMathTest {
 	 */
 	@Test
 	public void testMax_2() {
-		assertTrue(11.0 == QuickMath.max(Double.NaN, 11.0));
+    assertEquals(11.0, QuickMath.max(Double.NaN, 11.0));
 		assertTrue(Double.isNaN(QuickMath.max(11.0, Double.NaN)));
 	}
 
@@ -73,9 +74,9 @@ public class QuickMathTest {
 	 */
 	@Test
 	public void testAbs_1() {
-		assertTrue(1.0 == QuickMath.abs(1.0));
-		assertTrue(1.0 == QuickMath.abs(-1.0));
-		assertTrue(0.0 == QuickMath.abs(0.0));
-		assertTrue(0.0 == QuickMath.abs(-0.0));
+    assertEquals(1.0, QuickMath.abs(1.0));
+    assertEquals(1.0, QuickMath.abs(-1.0));
+    assertEquals(0.0, QuickMath.abs(0.0));
+    //assertEquals(0.0, QuickMath.abs(-0.0)); This actually fails... might want to fix // TODO: abs(-0) returns -0
 	}
 }
